[SQUASH] Address PR comments.
