+++ freqtrade/strategy/__init__.py
@@ -3,6 +3,7 @@ from freqtrade.exchange import (timeframe_to_minutes, timeframe_to_msecs, timefr
                                 timeframe_to_prev_date, timeframe_to_seconds)
 from freqtrade.strategy.hyper import (BooleanParameter, CategoricalParameter, DecimalParameter,
                                       IntParameter, RealParameter)
+from freqtrade.strategy.informative_decorator import informative
 from freqtrade.strategy.interface import IStrategy
-from freqtrade.strategy.strategy_helper import (informative, merge_informative_pair,
-                                                stoploss_from_absolute, stoploss_from_open)
+from freqtrade.strategy.strategy_helper import (merge_informative_pair, stoploss_from_absolute,
+                                                stoploss_from_open)
